+++ neuralop/data/transforms/patching_transforms.py
@@ -112,48 +112,3 @@ class MGPTensorDataset(Dataset):
 
     def __len__(self):
         return self.x.size(0)
-
-def regular_grid_2d(spatial_dims, grid_boundaries=[[0, 1], [0, 1]]):
-    """
-    Appends grid positional encoding to an input tensor, concatenating as additional dimensions along the channels
-    """
-    height, width = spatial_dims
-
-    xt = torch.linspace(grid_boundaries[0][0], grid_boundaries[0][1],
-                        height + 1)[:-1]
-    yt = torch.linspace(grid_boundaries[1][0], grid_boundaries[1][1],
-                        width + 1)[:-1]
-
-    grid_x, grid_y = torch.meshgrid(xt, yt, indexing='ij')
-
-    grid_x = grid_x.repeat(1, 1)
-    grid_y = grid_y.repeat(1, 1)
-
-    return grid_x, grid_y
-
-def regular_grid_nd(resolutions: List[int], grid_boundaries: List[List[int]]=[[0,1]] * 2):
-    """regular_grid_nd generates a tensor of coordinate points that 
-    describe a bounded regular grid.
-    
-    Parameters
-    ----------
-    resolutions : List[int]
-        resolution of the output grid along each dimension
-    grid_boundaries : List[List[int]], optional
-        List of pairs [start, end] of the boundaries of the
-        regular grid. Must correspond 1-to-1 with resolutions default [[0,1], [0,1]]
-
-    Returns
-    -------
-    grid: tuple(Tensor)
-    list of tensors describing positional encoding 
-    """
-    assert len(resolutions) == len(grid_boundaries), "Error: inputs must have same number of dimensions"
-    dim = len(resolutions)
-
-    meshgrid_inputs = list()
-    for res, (start,stop) in zip(resolutions, grid_boundaries):
-        meshgrid_inputs.append(torch.linspace(start, stop, res + 1)[:-1])
-    grid = torch.meshgrid(*meshgrid_inputs, indexing='ij')
-    grid = tuple([x.repeat([1]*dim) for x in grid])
-    return grid
