+++ neuralop/data/transforms/positional_embeddings.py
@@ -3,13 +3,11 @@ from typing import List
 import torch
 from torch import nn
 
-from .grid_transforms import regular_grid_2d, regular_grid_nd
-
-class Euclidean2D(nn.Module):
+class GridEmbedding2D(nn.Module):
     """A simple positional embedding as a regular 2D grid
     """
     def __init__(self, grid_boundaries=[[0, 1], [0, 1]]):
-        """Euclidean2D applies a simple positional 
+        """GridEmbedding2D applies a simple positional 
         embedding as a regular 2D grid
 
         Parameters
@@ -67,11 +65,11 @@ class Euclidean2D(nn.Module):
         else:
             return out
 
-class EuclideanND(nn.Module):
+class GridEmbeddingND(nn.Module):
     """A positional embedding as a regular ND grid
     """
     def __init__(self, dim: int=2, grid_boundaries=[[0, 1], [0, 1]]):
-        """EuclideanND applies a simple positional 
+        """GridEmbeddingND applies a simple positional 
         embedding as a regular ND grid
 
         Parameters
@@ -140,9 +138,9 @@ class EuclideanND(nn.Module):
                          dim=1)
         return out
     
-class Sinusoidal2D(nn.Module):
+class SinusoidalEmbedding2D(nn.Module):
     def __init__(self, num_channels, max_positions=10000, endpoint=False):
-        """Sinusoidal2D applies a 2d sinusoidal positional encoding 
+        """SinusoidalEmbedding2D applies a 2d sinusoidal positional encoding 
 
         Parameters
         ----------
@@ -167,31 +165,9 @@ class Sinusoidal2D(nn.Module):
         x = x.ger(freqs.to(x.dtype))
         x = torch.cat([x.cos(), x.sin()], dim=1)
         return x
-    
-# Utility fucntions for Rotary embedding
-# modified from https://github.com/lucidrains/x-transformers/blob/main/x_transformers/x_transformers.py
-def rotate_half(x):
-    """
-    Split x's channels into two equal halves.
-    """
-    # split the last dimension of x into two equal halves
-    x = x.reshape(*x.shape[:-1], 2, -1)
-    x1, x2 = x.unbind(dim=-2)
-    return torch.cat((-x2, x1), dim=-1)
 
 
-def apply_rotary_pos_emb(t, freqs):
-    """
-    Apply rotation matrix computed based on freqs to rotate t.
-    t: tensor of shape [batch_size, num_points, dim]
-    freqs: tensor of shape [batch_size, num_points, 1]
-
-    Formula: see equation (34) in https://arxiv.org/pdf/2104.09864.pdf
-    """
-    return (t * freqs.cos()) + (rotate_half(t) * freqs.sin())
-
-
-class Rotary(nn.Module):
+class RotaryEmbedding2D(nn.Module):
     def __init__(self, dim, min_freq=1/64, scale=1.):
         """
         Applying rotary positional embedding (https://arxiv.org/abs/2104.09864) to the input feature tensor.
@@ -222,3 +198,77 @@ class Rotary(nn.Module):
 
         return torch.cat((apply_rotary_pos_emb(t_x, freqs_x),
                           apply_rotary_pos_emb(t_y, freqs_y)), dim=-1)
+    
+
+# Utility functions for GridEmbedding
+def regular_grid_2d(spatial_dims, grid_boundaries=[[0, 1], [0, 1]]):
+    """
+    Creates a 2 x height x width stack of positional encodings A, where
+    A[:,i,j] = [[x,y]] at coordinate (i,j) on a (height, width) grid. 
+    """
+    height, width = spatial_dims
+
+    xt = torch.linspace(grid_boundaries[0][0], grid_boundaries[0][1],
+                        height + 1)[:-1]
+    yt = torch.linspace(grid_boundaries[1][0], grid_boundaries[1][1],
+                        width + 1)[:-1]
+
+    grid_x, grid_y = torch.meshgrid(xt, yt, indexing='ij')
+
+    grid_x = grid_x.repeat(1, 1)
+    grid_y = grid_y.repeat(1, 1)
+
+    return grid_x, grid_y
+
+def regular_grid_nd(resolutions: List[int], grid_boundaries: List[List[int]]=[[0,1]] * 2):
+    """regular_grid_nd generates a tensor of coordinate points that 
+    describe a bounded regular grid.
+    
+    Creates a dim x res_d1 x ... x res_dn stack of positional encodings A, where
+    A[:,c1,c2,...] = [[d1,d2,...dn]] at coordinate (c1,c2,...cn) on a (res_d1, ...res_dn) grid. 
+
+    Parameters
+    ----------
+    resolutions : List[int]
+        resolution of the output grid along each dimension
+    grid_boundaries : List[List[int]], optional
+        List of pairs [start, end] of the boundaries of the
+        regular grid. Must correspond 1-to-1 with resolutions default [[0,1], [0,1]]
+
+    Returns
+    -------
+    grid: tuple(Tensor)
+    list of tensors describing positional encoding 
+    """
+    assert len(resolutions) == len(grid_boundaries), "Error: inputs must have same number of dimensions"
+    dim = len(resolutions)
+
+    meshgrid_inputs = list()
+    for res, (start,stop) in zip(resolutions, grid_boundaries):
+        meshgrid_inputs.append(torch.linspace(start, stop, res + 1)[:-1])
+    grid = torch.meshgrid(*meshgrid_inputs, indexing='ij')
+    grid = tuple([x.repeat([1]*dim) for x in grid])
+    return grid
+
+  
+# Utility fucntions for Rotary embedding
+# modified from https://github.com/lucidrains/x-transformers/blob/main/x_transformers/x_transformers.py
+def rotate_half(x):
+    """
+    Split x's channels into two equal halves.
+    """
+    # split the last dimension of x into two equal halves
+    x = x.reshape(*x.shape[:-1], 2, -1)
+    x1, x2 = x.unbind(dim=-2)
+    return torch.cat((-x2, x1), dim=-1)
+
+
+def apply_rotary_pos_emb(t, freqs):
+    """
+    Apply rotation matrix computed based on freqs to rotate t.
+    t: tensor of shape [batch_size, num_points, dim]
+    freqs: tensor of shape [batch_size, num_points, 1]
+
+    Formula: see equation (34) in https://arxiv.org/pdf/2104.09864.pdf
+    """
+    return (t * freqs.cos()) + (rotate_half(t) * freqs.sin())
